cheri: fs: ecryptfs: Return allocated page(s) as uintptr_t
ecryptfs has its own wrapper to get a zeroed page. Return uintptr_t
instead of unsigned long from that wrapper.

+++ fs/ecryptfs/crypto.c
@@ -1019,7 +1019,7 @@ out:
 	return rc;
 }
 
-static unsigned long ecryptfs_get_zeroed_pages(gfp_t gfp_mask,
+static uintptr_t ecryptfs_get_zeroed_pages(gfp_t gfp_mask,
 					       unsigned int order)
 {
 	struct page *page;
